+++ src/api/cms/votations/services/users/remove.service.ts
@@ -4,7 +4,7 @@ import { votationsResponses } from "@/config/responses";
 import { RemoveVotationUsersBody } from "../../votations.schemas";
 
 export const removeVotationUsersService = orgFactory.createHandlers(async (c) => {
-  const { users } = await c.req.json<RemoveVotationUsersBody>();
+  const { data } = await c.req.json<RemoveVotationUsersBody>();
   const { id: organizationId } = c.get("orgData");
   const { votationId } = c.req.param();
 
@@ -13,11 +13,19 @@ export const removeVotationUsersService = orgFactory.createHandlers(async (c) =>
   const fetchedVotation = await votations.findById(organizationId, votationId);
   if (!fetchedVotation) return c.json(votationsResponses.notExists, 404);
 
-  for (const userId of users) {
+  for (const userId of data) {
     const existingUser = await votations.users.findById(organizationId, votationId, userId);
     if (existingUser) usersToRemove.push({ organizationId, votationId, userId });
   }
 
-  await votations.users.deleteMany(usersToRemove);
-  return c.json(votationsResponses.users.removed, 201);
+  if (usersToRemove.length) {
+    await votations.users.deleteMany(usersToRemove);
+  }
+
+  return c.json(
+    usersToRemove.length === data.length
+      ? votationsResponses.users.removed
+      : votationsResponses.users.removedPartially,
+    200,
+  );
 });
